pyproject.toml
[project]
name = "voice-transcriber-mcp"
version = "2.0.0"
description = "MCP server for transcribing voice messages from Google Chat using Groq Whisper API"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.0.0",
    "requests>=2.28.0",
]

[project.scripts]
voice-transcriber = "server:main"
